feat(docdb): honor Filters on the Describes that document them
DocumentDB accepted `Filters` and ignored it, so a caller narrowing a
listing got every resource in the account back -- the same bug fixed for
RDS in #2492.

- DescribeDBClusters (db-cluster-id), DescribeDBInstances
  (db-cluster-id, db-instance-id) and DescribeGlobalClusters
  (db-cluster-id) filter as AWS documents. Each accepts identifiers and
  ARNs; the instance path rebuilds its cluster's ARN from its own, which
  shares the partition, region and account.
- Filtering a global cluster by a MEMBER cluster's ARN selects the
  global cluster containing it, which is how a caller holding a regional
  ARN finds its global parent.
- Filters AND with each other and with the operation's own identifier
  parameter. An unrecognized name matches nothing and logs at warn:
  DocumentDB declares no InvalidParameterValue-equivalent on these
  operations, so rejecting would put an undeclared shape on the wire.

The parsing moved to fakecloud-core::query_filters, where Neptune can
reach it too -- RDS, DocumentDB and Neptune are all Query-protocol
services taking the same member with the same wire form, and all three
name resources with arn:aws:rds: ARNs, so the identifier helpers apply
unchanged. RDS keeps a re-export under the name its call sites already
use; the 16 parsing tests moved with the code.

Verified the filters bite by neutering the cluster predicate and
watching the test fail. docdb conformance holds at 1513/1513.
