+++ crates/fakecloud-core/src/query_filters.rs
@@ -1,0 +1,709 @@
+//! Parsing and evaluation of the `Filters` request member on the
+//! RDS-family Query APIs.
+//!
+//! RDS, DocumentDB and Neptune share one shape here -- all three are
+//! Query-protocol services whose Describe* operations take a
+//! `FilterList`, and all three name their resources with `arn:aws:rds:`
+//! ARNs, so the identifier helpers below apply unchanged to each.
+//!
+//! Their Describe* operations take a `FilterList` whose Smithy shape names
+//! the list member `Filter` and the value member `Value`, so the
+//! query-protocol wire form is:
+//!
+//! ```text
+//! Filters.Filter.1.Name=dbi-resource-id
+//! Filters.Filter.1.Values.Value.1=db-5e1b7794bef34787abd8310859cc93a2
+//! Filters.Filter.1.Values.Value.2=db-1b8e439e59564e849b38b04191c40b8e
+//! ```
+//!
+//! Older SDKs (and hand-rolled clients) sometimes emit the generic
+//! `member` element name instead, so both spellings are accepted.
+//!
+//! Semantics, per the AWS docs: filters are AND-ed with each other and
+//! with the operation's own identifier parameter; the values inside one
+//! filter are OR-ed. Names are case-sensitive, and so are values except on the few
+//! enum-valued filters AWS documents in a different case than it
+//! returns (see `matches_ignore_case`), and wildcards
+//! are not supported.
+//!
+//! Real RDS rejects a filter name an operation doesn't support with
+//! `InvalidParameterValue`. We can't: that error is not declared on any
+//! of the Describe* operations in the Smithy model, so returning it
+//! would emit an undeclared error shape and fail conformance. The
+//! closest in-shape behaviour is to treat an unrecognized filter (and a
+//! filter carrying no values) as matching no resource, which is what the
+//! per-operation matchers below do — a caller filtering on something we
+//! don't recognise gets an empty result rather than the whole list.
+
+use std::collections::{HashMap, HashSet};
+
+use crate::service::AwsRequest;
+
+/// Look a parameter up by presence, keeping an explicitly-empty value.
+///
+/// `optional_query_param` treats `Key=` as absent, which would truncate a
+/// filter's value list at a blank member (`Values.Value.1=` followed by
+/// `Values.Value.2=mysql` would parse as no values at all, and the filter
+/// would then match nothing). AWS keeps the non-empty siblings, so the
+/// list walk has to distinguish "present but empty" from "absent".
+///
+/// Dispatch merges the form body into `query_params` for the Query
+/// protocol, so `body` is normally empty here; it is consulted as a
+/// fallback (parsed once by the caller) so a request built with only a
+/// body -- an in-process caller, a future entry point -- can't silently
+/// see zero filters and get the whole unfiltered list back.
+fn present_param<'a>(
+    req: &'a AwsRequest,
+    body: &'a HashMap<String, String>,
+    key: &str,
+) -> Option<String> {
+    req.query_params.get(key).or_else(|| body.get(key)).cloned()
+}
+
+/// One `Filters.Filter.N` entry: a name plus the values it accepts.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct QueryFilter {
+    pub name: String,
+    pub values: Vec<String>,
+}
+
+impl QueryFilter {
+    /// True when `candidate` is one of the filter's values. A `None`
+    /// candidate (the resource doesn't carry the attribute at all) never
+    /// matches, which is what AWS does for e.g. `domain` on an instance
+    /// with no Active Directory membership.
+    pub fn matches(&self, candidate: Option<&str>) -> bool {
+        match candidate {
+            Some(value) => self.values.iter().any(|v| v == value),
+            None => false,
+        }
+    }
+
+    /// True when `candidate` is one of the filter's values, compared
+    /// case-insensitively.
+    ///
+    /// Filters are case-sensitive on AWS, but a few enum-valued ones
+    /// are documented in a different case than the API stores and
+    /// returns -- `db-cluster-endpoint-type` accepts `reader` while the
+    /// endpoint reads back `READER`. An exact comparison there means a
+    /// caller copying the documented command selects nothing.
+    pub fn matches_ignore_case(&self, candidate: Option<&str>) -> bool {
+        match candidate {
+            Some(value) => self.values.iter().any(|v| v.eq_ignore_ascii_case(value)),
+            None => false,
+        }
+    }
+
+    /// True when any of `candidates` is one of the filter's values. Used
+    /// for filters AWS documents as accepting "identifiers and ARNs".
+    pub fn matches_any<'a>(&self, candidates: impl IntoIterator<Item = Option<&'a str>>) -> bool {
+        candidates.into_iter().any(|c| self.matches(c))
+    }
+}
+
+/// Reports the filter names an operation doesn't support, once per
+/// request.
+///
+/// Called before the rows are scanned, not from inside the per-row
+/// predicate: reporting from there logged one line per row, and a cache
+/// to dedupe it would grow without bound on request-supplied names.
+/// Here the work is bounded by the request's own filter count and the
+/// predicate stays side-effect-free.
+pub fn warn_unknown_filters(filters: &[QueryFilter], supported: &[&str]) {
+    // Deduped within the request: a caller may repeat the same name
+    // across entries, and one line per entry is the same noise the
+    // per-row report made, just smaller. A local set keeps this bounded
+    // by the request rather than by anything the caller can grow, and
+    // keeps the pass linear -- the names are request-controlled, so a
+    // scan per name would let a long list cost quadratic time before
+    // the Describe even starts.
+    let mut reported: HashSet<&str> = HashSet::new();
+    for filter in filters {
+        let name = filter.name.as_str();
+        if supported.contains(&name) || !reported.insert(name) {
+            continue;
+        }
+        tracing::warn!(
+            filter = %name,
+            "unrecognized RDS filter name; matching no resource"
+        );
+    }
+}
+
+/// The identifier a Describe was asked to narrow to, or `None` when it
+/// was not asked to narrow at all.
+///
+/// `normalized_identifier` alone is not enough here: it returns `None`
+/// both for "absent" and for "an ARN this operation can't resolve"
+/// (wrong resource type, or another account's), and a caller reading
+/// `None` as "no narrowing" would answer a targeted request with the
+/// WHOLE list -- and skip the not-found check on the way. Failing
+/// closed, an unresolvable identifier stays as the caller wrote it: it
+/// matches no stored row, so the request selects nothing or raises the
+/// operation's declared not-found fault.
+pub fn requested_identifier(
+    raw: Option<String>,
+    resource_type: &str,
+    account_id: &str,
+) -> Option<String> {
+    // An explicitly empty value is not a request for a resource named
+    // "" -- every other parameter here treats it as absent.
+    let raw = raw.filter(|value| !value.is_empty())?;
+    if !addresses_own_account(&raw, account_id) {
+        // Another account's ARN must not alias onto this account's
+        // same-named resource.
+        return Some(raw);
+    }
+    Some(normalized_identifier(Some(raw.clone()), resource_type).unwrap_or(raw))
+}
+
+/// Rebuild an RDS ARN for a sibling resource of `arn` (same partition,
+/// region and account) with a different resource type and id, so filters
+/// AWS documents as accepting "identifiers and ARNs" can be matched
+/// against the ARN form too. Returns `None` for a malformed ARN.
+pub fn sibling_rds_arn(arn: &str, resource_type: &str, id: &str) -> Option<String> {
+    // Guarded on the `arn:` prefix AND the `rds` service, like
+    // `identifier_matches_type`: otherwise any 5-field string -- or an
+    // ARN of another service -- would be turned into a fabricated ARN
+    // that a filter could then match on.
+    if !arn.starts_with("arn:") {
+        return None;
+    }
+    let prefix: Vec<&str> = arn.splitn(7, ':').take(5).collect();
+    if prefix.len() < 5 || prefix[2] != "rds" {
+        return None;
+    }
+    Some(format!("{}:{resource_type}:{id}", prefix.join(":")))
+}
+
+/// Normalize an identifier request parameter: an explicitly-empty value
+/// means "absent" (AWS ignores it rather than matching the empty string),
+/// and an ARN of `resource_type` is reduced to its resource segment
+/// because clients pass either form — the Terraform provider stores full
+/// ARNs in `snapshot_identifier`. An ARN of any OTHER type resolves to
+/// nothing rather than aliasing onto a same-named resource of the type
+/// the caller actually asked about.
+///
+/// The reduction is guarded on the `arn:` prefix, and takes everything
+/// after the resource-type field rather than the last colon segment:
+/// AWS's automated-snapshot identifiers themselves contain a colon
+/// (`rds:mydb-2026-08-30-06-00`), both bare and inside an ARN
+/// (`arn:aws:rds:us-east-1:123456789012:snapshot:rds:mydb-...`), so
+/// either kind of blind trimming turns a real id into a lookup miss.
+///
+/// Note this is the opposite of a filter *value*, where an explicit empty
+/// string is a legitimate member to match on (see `present_param`).
+pub fn normalized_identifier(param: Option<String>, resource_type: &str) -> Option<String> {
+    param
+        .filter(|value| !value.is_empty())
+        .and_then(|value| {
+            if !value.starts_with("arn:") {
+                return Some(value);
+            }
+            // arn:partition:service:region:account:type:id -- the type
+            // field has to match, or a cluster ARN would address a DB
+            // instance of the same name; `id` is field 7 and keeps any
+            // colons of its own.
+            if !identifier_matches_type(&value, resource_type) {
+                return None;
+            }
+            value.splitn(7, ':').nth(6).map(str::to_string)
+        })
+        .filter(|value| !value.is_empty())
+}
+
+/// Parse a request flag leniently.
+///
+/// `parse_optional_bool` accepts only `true`/`false`, and these flags
+/// can't reject a junk value (`InvalidParameterValue` is undeclared on
+/// the Describe ops), so an unrecognized value would silently read as
+/// `false` -- a wrong answer rather than a tolerated one. `1`/`0` are
+/// honoured too; anything else is treated as absent.
+pub fn optional_flag(value: Option<&str>) -> Option<bool> {
+    match value?.trim() {
+        v if v.eq_ignore_ascii_case("true") || v == "1" => Some(true),
+        v if v.eq_ignore_ascii_case("false") || v == "0" => Some(false),
+        _ => None,
+    }
+}
+
+/// True when an identifier is a bare id, or an ARN of `resource_type`.
+///
+/// [`normalized_identifier`] answers `None` both for "parameter absent"
+/// and for "ARN of the wrong resource type", and every Describe treats a
+/// `None` identifier as "no filter" -- so a wrong-type ARN would widen a
+/// targeted read into a full listing. Callers check this first and raise
+/// the operation's not-found instead.
+pub fn identifier_matches_type(param: &str, resource_type: &str) -> bool {
+    if !param.starts_with("arn:") {
+        return true;
+    }
+    let mut fields = param.splitn(7, ':').skip(2);
+    // The service has to be `rds` (an `arn:aws:ec2:...:db:mydb` names
+    // nothing here), the ACCOUNT has to be non-empty (an ARN naming no
+    // account resolves to nothing rather than aliasing onto the caller's
+    // own resource -- `identifier_account` reports None for it, which
+    // the resolve paths cannot tell from a bare id), the type has to
+    // match, AND the resource id has to be non-empty:
+    // `arn:aws:rds:us-east-1:1234:db:` would otherwise pass, normalize
+    // to `None`, and read as "no identifier" -- widening a targeted read
+    // into a full listing.
+    fields.next() == Some("rds")
+        // Region and account both have to be real: RDS ARNs always carry
+        // them, and an empty field would let `arn:aws:rds::1234:db:mydb`
+        // resolve against this account's instance whatever the request
+        // region.
+        && fields.next().is_some_and(|region| !region.is_empty())
+        && fields.next().is_some_and(|account| !account.is_empty())
+        && fields.next() == Some(resource_type)
+        && fields.next().is_some_and(|id| !id.is_empty())
+}
+
+/// The account an ARN-form identifier names, or `None` for a bare id.
+///
+/// [`normalized_identifier`] reduces an ARN to its resource id, which is
+/// what the per-account state is keyed on -- but the account field can't
+/// just be dropped: with cross-account sharing, account B passing account
+/// A's full ARN must not silently resolve to B's own same-named resource
+/// (on `DeleteDBSnapshot` that would delete the wrong snapshot).
+pub fn identifier_account(param: &str) -> Option<String> {
+    if !param.starts_with("arn:") {
+        return None;
+    }
+    param
+        .split(':')
+        .nth(4)
+        .map(str::to_string)
+        .filter(|account| !account.is_empty())
+}
+
+/// True when an identifier addresses a resource the caller owns: a bare
+/// id, or an ARN naming the caller's own account. Operations that only
+/// act on owned resources (delete, copy source, attribute changes) use
+/// this to reject another account's ARN instead of aliasing it onto a
+/// same-named local resource.
+pub fn addresses_own_account(param: &str, account_id: &str) -> bool {
+    match identifier_account(param) {
+        Some(account) => account == account_id,
+        // A bare id is this account's. An ARN with an EMPTY account
+        // field names nothing resolvable -- treating it as a bare id
+        // would let `arn:aws:rds:us-east-1::pg:mypg` address the local
+        // resource of that name.
+        None => !param.starts_with("arn:"),
+    }
+}
+
+/// Parse `Filters.Filter.N.Name` + `Filters.Filter.N.Values.Value.M` (and
+/// the `member` spelling of either element) into filter entries.
+///
+/// Indices are 1-based and contiguous; parsing stops at the first gap,
+/// matching how the SDKs serialize the list.
+pub fn parse_filters(req: &AwsRequest) -> Vec<QueryFilter> {
+    let mut filters = Vec::new();
+    // Parsed once, not per lookup, and consulted whenever a body is
+    // present -- `extras::get_param` falls back unconditionally, and
+    // gating on an empty `query_params` would miss a request that
+    // carries `Action` in the query string and the rest in the body.
+    // Dispatch merges the form body into `query_params` for the Query
+    // protocol, so the fallback is only for a request built without that
+    // merge. Parsing the body on every Describe just to service a path
+    // production never takes would allocate a map of every request
+    // parameter on a hot path -- and every unfiltered Describe carries a
+    // body, so "the body is non-empty" is not a discriminating gate. Scan
+    // the raw bytes for the parameter prefix first: a substring search
+    // allocates nothing, and it decides the common case (no filters
+    // anywhere) without building the map. `Filters.` also can't appear in
+    // a Query request except as a parameter name -- values are
+    // percent-encoded, so a `.` in one arrives as `%2E`.
+    let filters_in_query = req
+        .query_params
+        .keys()
+        .any(|key| key.starts_with("Filters."));
+    const FILTER_PREFIX: &[u8] = b"Filters.";
+    let filters_in_body = !filters_in_query
+        && req
+            .body
+            .windows(FILTER_PREFIX.len())
+            .any(|window| window == FILTER_PREFIX);
+    let body = if filters_in_body {
+        crate::protocol::parse_query_body(&req.body)
+    } else {
+        HashMap::new()
+    };
+
+    for index in 1.. {
+        let Some((prefix, name)) = ["Filter", "member"].iter().find_map(|element| {
+            let prefix = format!("Filters.{element}.{index}");
+            present_param(req, &body, &format!("{prefix}.Name")).map(|name| (prefix, name))
+        }) else {
+            break;
+        };
+
+        let mut values = Vec::new();
+        for element in ["Value", "member"] {
+            for value_index in 1.. {
+                let key = format!("{prefix}.Values.{element}.{value_index}");
+                match present_param(req, &body, &key) {
+                    Some(value) => values.push(value),
+                    None => break,
+                }
+            }
+            if !values.is_empty() {
+                break;
+            }
+        }
+
+        filters.push(QueryFilter { name, values });
+    }
+
+    filters
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use bytes::Bytes;
+    use http::{HeaderMap, Method};
+    use std::collections::HashMap;
+
+    fn request(params: &[(&str, &str)]) -> AwsRequest {
+        let mut query_params =
+            HashMap::from([("Action".to_string(), "DescribeDBInstances".to_string())]);
+        for (key, value) in params {
+            query_params.insert((*key).to_string(), (*value).to_string());
+        }
+        AwsRequest {
+            service: "rds".to_string(),
+            action: "DescribeDBInstances".to_string(),
+            region: "us-east-1".to_string(),
+            account_id: "123456789012".to_string(),
+            request_id: "test-request-id".to_string(),
+            headers: HeaderMap::new(),
+            query_params,
+            body: Bytes::new(),
+            body_stream: parking_lot::Mutex::new(None),
+            path_segments: vec![],
+            raw_path: "/".to_string(),
+            raw_query: String::new(),
+            method: Method::POST,
+            is_query_protocol: true,
+            access_key_id: None,
+            principal: None,
+        }
+    }
+
+    /// The body fallback exists for a caller that puts the parameters in
+    /// the form body without dispatch's merge -- including one that keeps
+    /// `Action` in the query string, so an "empty query_params" gate
+    /// would miss it. And an ordinary unfiltered Describe must not pay
+    /// for a parse it has no use for.
+    #[test]
+    fn reads_filters_from_an_unmerged_form_body() {
+        let mut req = request(&[]);
+        req.body = Bytes::from_static(
+            b"Action=DescribeDBInstances&Filters.Filter.1.Name=engine&Filters.Filter.1.Values.Value.1=mysql",
+        );
+
+        assert_eq!(
+            parse_filters(&req),
+            vec![QueryFilter {
+                name: "engine".to_string(),
+                values: vec!["mysql".to_string()],
+            }]
+        );
+
+        // A body with no filter parameter at all yields none -- the
+        // common path, and the one the byte scan short-circuits.
+        let mut req = request(&[]);
+        req.body = Bytes::from_static(b"Action=DescribeDBInstances&MaxRecords=20");
+        assert!(parse_filters(&req).is_empty());
+
+        // A filter already visible in the query wins; the body is not
+        // consulted, so a stale duplicate there can't double it up.
+        let mut req = request(&[
+            ("Filters.Filter.1.Name", "engine"),
+            ("Filters.Filter.1.Values.Value.1", "postgres"),
+        ]);
+        req.body = Bytes::from_static(
+            b"Filters.Filter.1.Name=engine&Filters.Filter.1.Values.Value.1=mysql",
+        );
+        assert_eq!(
+            parse_filters(&req),
+            vec![QueryFilter {
+                name: "engine".to_string(),
+                values: vec!["postgres".to_string()],
+            }]
+        );
+    }
+
+    #[test]
+    fn parses_filter_element_spelling() {
+        let filters = parse_filters(&request(&[
+            ("Filters.Filter.1.Name", "dbi-resource-id"),
+            ("Filters.Filter.1.Values.Value.1", "db-a"),
+            ("Filters.Filter.1.Values.Value.2", "db-b"),
+            ("Filters.Filter.2.Name", "engine"),
+            ("Filters.Filter.2.Values.Value.1", "mysql"),
+        ]));
+
+        assert_eq!(
+            filters,
+            vec![
+                QueryFilter {
+                    name: "dbi-resource-id".to_string(),
+                    values: vec!["db-a".to_string(), "db-b".to_string()],
+                },
+                QueryFilter {
+                    name: "engine".to_string(),
+                    values: vec!["mysql".to_string()],
+                },
+            ]
+        );
+    }
+
+    #[test]
+    fn parses_member_element_spelling() {
+        let filters = parse_filters(&request(&[
+            ("Filters.member.1.Name", "engine"),
+            ("Filters.member.1.Values.member.1", "postgres"),
+        ]));
+
+        assert_eq!(
+            filters,
+            vec![QueryFilter {
+                name: "engine".to_string(),
+                values: vec!["postgres".to_string()],
+            }]
+        );
+    }
+
+    #[test]
+    fn keeps_an_explicitly_empty_value() {
+        // `Value.1=` is a legitimate member (filtering for a blank
+        // attribute); dropping it must not truncate the list and lose
+        // `Value.2`.
+        let filters = parse_filters(&request(&[
+            ("Filters.Filter.1.Name", "engine"),
+            ("Filters.Filter.1.Values.Value.1", ""),
+            ("Filters.Filter.1.Values.Value.2", "mysql"),
+        ]));
+
+        assert_eq!(
+            filters,
+            vec![QueryFilter {
+                name: "engine".to_string(),
+                values: vec![String::new(), "mysql".to_string()],
+            }]
+        );
+    }
+
+    #[test]
+    fn parses_filters_from_a_form_body() {
+        // Dispatch normally merges the body into `query_params`; a
+        // request carrying only a body must not silently parse as "no
+        // filters" and return the whole unfiltered list.
+        // `Action` stays in the query string, the filters only in the
+        // body -- the shape a gated fallback would miss.
+        let mut req = request(&[]);
+        req.body = Bytes::from(
+            "Action=DescribeDBInstances&Filters.Filter.1.Name=dbi-resource-id\
+             &Filters.Filter.1.Values.Value.1=db-a",
+        );
+
+        assert_eq!(
+            parse_filters(&req),
+            vec![QueryFilter {
+                name: "dbi-resource-id".to_string(),
+                values: vec!["db-a".to_string()],
+            }]
+        );
+    }
+
+    #[test]
+    fn stops_at_first_index_gap() {
+        let filters = parse_filters(&request(&[
+            ("Filters.Filter.1.Name", "engine"),
+            ("Filters.Filter.1.Values.Value.1", "mysql"),
+            ("Filters.Filter.3.Name", "domain"),
+            ("Filters.Filter.3.Values.Value.1", "d-1"),
+        ]));
+
+        assert_eq!(filters.len(), 1);
+    }
+
+    #[test]
+    fn absent_filters_parse_to_empty() {
+        assert!(parse_filters(&request(&[("MaxRecords", "20")])).is_empty());
+    }
+
+    #[test]
+    fn filter_without_values_matches_nothing() {
+        let filter = QueryFilter {
+            name: "engine".to_string(),
+            values: Vec::new(),
+        };
+
+        assert!(!filter.matches(Some("mysql")));
+    }
+
+    #[test]
+    fn matching_is_case_sensitive_and_ors_values() {
+        let filter = QueryFilter {
+            name: "engine".to_string(),
+            values: vec!["mysql".to_string(), "postgres".to_string()],
+        };
+
+        assert!(filter.matches(Some("mysql")));
+        assert!(filter.matches(Some("postgres")));
+        assert!(!filter.matches(Some("MySQL")));
+        assert!(!filter.matches(None));
+    }
+
+    #[test]
+    fn matches_any_accepts_alternate_attributes() {
+        let filter = QueryFilter {
+            name: "db-instance-id".to_string(),
+            values: vec!["arn:aws:rds:us-east-1:000000000000:db:mydb".to_string()],
+        };
+
+        assert!(filter.matches_any([
+            Some("mydb-other"),
+            Some("arn:aws:rds:us-east-1:000000000000:db:mydb"),
+        ]));
+        assert!(!filter.matches_any([Some("mydb-other"), None]));
+    }
+
+    #[test]
+    fn normalized_identifier_drops_empty_and_reduces_arns() {
+        assert_eq!(normalized_identifier(None, "snapshot"), None);
+        // `Key=` reaches handlers as Some("") and means "not supplied".
+        assert_eq!(normalized_identifier(Some(String::new()), "snapshot"), None);
+        assert_eq!(
+            normalized_identifier(Some("snap-1".to_string()), "snapshot"),
+            Some("snap-1".to_string())
+        );
+        assert_eq!(
+            normalized_identifier(
+                Some("arn:aws:rds:us-east-1:123456789012:snapshot:snap-1".to_string()),
+                "snapshot"
+            ),
+            Some("snap-1".to_string())
+        );
+        // AWS's automated-snapshot ids carry a colon and are NOT ARNs;
+        // trimming at the last colon would turn a real id into a miss.
+        assert_eq!(
+            normalized_identifier(Some("rds:mydb-2026-08-30-06-00".to_string()), "snapshot"),
+            Some("rds:mydb-2026-08-30-06-00".to_string())
+        );
+        // ...and the id keeps its colon inside an ARN too, so the
+        // reduction takes the whole resource field, not the last segment.
+        assert_eq!(
+            normalized_identifier(
+                Some(
+                    "arn:aws:rds:us-east-1:123456789012:snapshot:rds:mydb-2026-08-30-06-00"
+                        .to_string()
+                ),
+                "snapshot"
+            ),
+            Some("rds:mydb-2026-08-30-06-00".to_string())
+        );
+        // An ARN of a DIFFERENT resource type resolves to nothing rather
+        // than aliasing onto a same-named resource of the asked-for type.
+        assert_eq!(
+            normalized_identifier(
+                Some("arn:aws:rds:us-east-1:123456789012:cluster:foo".to_string()),
+                "db"
+            ),
+            None
+        );
+        assert_eq!(
+            normalized_identifier(
+                Some("arn:aws:rds:us-east-1:123456789012:cluster-snapshot:snap-1".to_string()),
+                "snapshot"
+            ),
+            None
+        );
+    }
+
+    #[test]
+    fn optional_flag_accepts_the_forms_clients_send() {
+        assert_eq!(optional_flag(Some("true")), Some(true));
+        assert_eq!(optional_flag(Some("TRUE")), Some(true));
+        assert_eq!(optional_flag(Some("1")), Some(true));
+        assert_eq!(optional_flag(Some("false")), Some(false));
+        assert_eq!(optional_flag(Some("0")), Some(false));
+        // Junk is absent, not false.
+        assert_eq!(optional_flag(Some("yes-please")), None);
+        assert_eq!(optional_flag(None), None);
+    }
+
+    #[test]
+    fn identifier_matches_type_guards_the_resource_field() {
+        assert!(identifier_matches_type("snap-1", "snapshot"));
+        assert!(identifier_matches_type(
+            "arn:aws:rds:us-east-1:123456789012:snapshot:snap-1",
+            "snapshot"
+        ));
+        assert!(!identifier_matches_type(
+            "arn:aws:rds:us-east-1:123456789012:cluster:mycl",
+            "db"
+        ));
+        // Right type, empty resource id: still not an identifier.
+        assert!(!identifier_matches_type(
+            "arn:aws:rds:us-east-1:123456789012:db:",
+            "db"
+        ));
+        assert!(!identifier_matches_type(
+            "arn:aws:rds:us-east-1:123456789012:db",
+            "db"
+        ));
+    }
+
+    #[test]
+    fn identifier_account_reads_the_arn_owner() {
+        assert_eq!(
+            identifier_account("arn:aws:rds:us-east-1:111111111111:snapshot:snap-1").as_deref(),
+            Some("111111111111")
+        );
+        assert_eq!(identifier_account("snap-1"), None);
+        assert!(addresses_own_account("snap-1", "222222222222"));
+        assert!(addresses_own_account(
+            "arn:aws:rds:us-east-1:222222222222:snapshot:snap-1",
+            "222222222222"
+        ));
+        // Another account's ARN must not alias onto a local resource.
+        assert!(!addresses_own_account(
+            "arn:aws:rds:us-east-1:111111111111:snapshot:snap-1",
+            "222222222222"
+        ));
+        // Nor an ARN whose account field is empty.
+        assert!(!addresses_own_account(
+            "arn:aws:rds:us-east-1::snapshot:snap-1",
+            "222222222222"
+        ));
+    }
+
+    #[test]
+    fn sibling_arn_swaps_resource_type_and_id() {
+        assert_eq!(
+            sibling_rds_arn(
+                "arn:aws:rds:us-east-1:123456789012:db:mydb",
+                "cluster",
+                "myclu"
+            ),
+            Some("arn:aws:rds:us-east-1:123456789012:cluster:myclu".to_string())
+        );
+    }
+
+    #[test]
+    fn sibling_arn_rejects_malformed_input() {
+        assert_eq!(sibling_rds_arn("not-an-arn", "db", "mydb"), None);
+        // A 5-field non-ARN must not be turned into a fabricated ARN.
+        assert_eq!(sibling_rds_arn("a:b:c:d:e", "cluster", "x"), None);
+        // Nor an ARN of another service.
+        assert_eq!(
+            sibling_rds_arn("arn:aws:ec2:us-east-1:123:instance:i-1", "cluster", "x"),
+            None
+        );
+    }
+}
